+++ iOS/Views/HomeworkSubmitFileFooterView.swift
@@ -1,0 +1,10 @@
+//
+//  Created for schulcloud-mobile-ios under GPL-3.0 license.
+//  Copyright © HPI. All rights reserved.
+//
+
+import UIKit
+
+final class HomeworkSubmitFileFooterView: UITableViewHeaderFooterView {
+    @IBOutlet var submitButton: UIButton!
+}
